refactor(api): claude_sessions.py 분할 완료 — processes·summaries 추출, _legacy 소멸
processes 3개를 processes.py로, summaries 2개를 summaries.py로 이동하고
남은 _legacy.py(빈 경로 라우트 2개)를 core.py로 rename 했다.

1,352줄 1파일 → 8모듈 최대 452줄. 25개 라우트 HTTP 표면 동일.

**슬라이스 1줄 누락 버그를 잡아 수정했다.** summaries.generate_batch_summaries
의 마지막 return results 가 추출 범위 밖이었다. 라인 범위 diff 검증이
같은 잘못된 범위를 양쪽에 써서 자기 확인이 됐고(IDENTICAL 오판), mypy 의
Missing return statement 만이 이를 잡았다 — 이 엔드포인트는 테스트가 0건이다.

검증을 라인 산술에서 AST 대조로 교체했다: 원본(4c85a37)과 분할 결과의
최상위 def/class 35개를 이름별로 매칭해 전부 바이트 동일 확인, 모듈 레벨
할당 7개(SortField·SortOrder·_line_count_cache·_sync_cache·_sync_lock·
logger·router) 유실 0건 확인. 중복 정의 0건.
