+++ src/backend/api/claude_sessions/__init__.py
@@ -3,6 +3,17 @@
 `api/claude_sessions.py`(1,352줄)를 도메인별 모듈로 분할한 결과. 소비자의
 import 경로는 분할 전과 동일하게 유지된다.
 
+| 모듈 | 책임 | 라우트 |
+|---|---|---:|
+| `core`       | 세션 컬렉션 + 집계 라우터 소유 (빈 경로 라우트) | 2 |
+| `sources`    | 외부 projects 경로 · 소스 사용자                | 4 |
+| `discovery`  | 프로젝트 목록 · 빈/유령 세션 조회·삭제          | 4 |
+| `processes`  | 로컬 claude OS 프로세스 관리                    | 3 |
+| `summaries`  | 요약 배치 생성·집계                             | 2 |
+| `sessions`   | 개별 세션 (`/{session_id}`)                     | 7 |
+| `activity`   | 대시보드 활동·태스크                            | 3 |
+| `sync`       | 백그라운드 DB 동기화 (라우트 없음)              | — |
+
 재노출 대상은 두 이름이다 (실측 2026-08-08):
 
 - `router`                          `api/app.py:90` 이 `safe_import` 로 가져가
@@ -19,8 +30,8 @@ import 경로는 분할 전과 동일하게 유지된다.
 그 테스트는 실제 바인딩 모듈을 직접 패치한다.
 """
 
-from . import activity, discovery, sessions, sources
-from ._legacy import router
+from . import activity, discovery, processes, sessions, sources, summaries
+from .core import router
 from .sync import scan_and_sync_claude_snapshots
 
 # **include 순서가 계약이다.** Starlette 는 등록 순서대로 전체 경로를 매칭하고
@@ -33,6 +44,8 @@ from .sync import scan_and_sync_claude_snapshots
 # 도달 불가가 된다. test_no_shadowing_route_pairs 가 이를 잡는다.
 router.include_router(sources.router)
 router.include_router(discovery.router)
+router.include_router(processes.router)
+router.include_router(summaries.router)
 router.include_router(sessions.router)
 router.include_router(activity.router)
